@@ -1,25 +1,23 @@
 {
   "name": "wsp001",
   "version": "1.0.0",
-  "description": "",
   "main": "01-fetch.js",
   "scripts": {
-    "start": "ts-node 01-data-demo.ts",
+    "start": "ts-node 01-data.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@types/express": "^4.17.13",
-    "@types/node": "^18.7.12",
+    "@types/node": "^18.7.13",
     "express": "^4.18.1",
     "node-fetch": "^3.2.10",
     "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
+    "typescript": "^4.8.2"
   },
   "devDependencies": {
     "jest": "^28.1.3"
-  }
+  },
+  "description": ""
 }
-
-
