@@ -1,10 +1,18 @@
+console.log("=============================1==========================");
+
 let a: number | string = 123;
 
 a = "dennis";
 
 console.log(a);
 
-let numArray: Array<string | number> = ["sdsafds", 1, 2, 3];
+console.log("=============================2==========================");
+
+// let numArray1 = ["dsfsd", 1, 2, 3, 4, [5], [[345]]];
+
+// let numArray2: Array<string | number> = ["sdsafds", 1, 2, 3];
+
+console.log("=============================3==========================");
 
 function formatDate(date: Date | string): String {
   let newDate = new Date(date);
@@ -19,8 +27,10 @@ console.log(formatDate(new Date()));
 console.log(formatDate("2019-07-21"));
 
 console.log(new Date());
-console.log(new Date().toLocaleString);
-////////////////////////////////////////////////////////////////////////////////////
+console.log(new Date().toLocaleString());
+// ////////////////////////////////////////////////////////////////////////////////////
+
+console.log("=============================4==========================");
 
 type Teacher = {
   alias: string;
@@ -43,8 +53,8 @@ const peter: Teacher = {
   alias: "Peter",
   age: 50,
   students: [
-    { name: "Andy", age: 20 },
-    { name: "Bob", age: 23 },
+    // { name: "Andy", age: 20 },
+    // { name: "Bob", age: 23 },
     {
       name: "Charlie",
       age: 25,
@@ -52,6 +62,8 @@ const peter: Teacher = {
     },
   ],
 };
+
+console.log("=============================5==========================");
 
 function greetStudent(student: Student) {
   console.log("hi", student.name);
